Properly integrated Teensy ethernet library while maintaning the same file structure by utilizing conditional compilation directives.
