+++ src/RoveCommEthernet/RoveCommEthernet.cpp
@@ -3,7 +3,6 @@
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////
 #include          <SPI.h>         // Energia/master/hardware/lm4f/libraries/SPI
-#include          <Ethernet.h>
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 void RoveCommEthernet::begin(const uint8_t ip_octet_4, EthernetServer* TCPServer)
