+++ src/RoveCommEthernetUDP/RoveCommEthernetUdp.h
@@ -3,8 +3,14 @@
 
 #include <stdint.h>
 #include <stddef.h>
+
+#if defined(ENERGIA)
 #include <Ethernet.h>
 #include <EthernetUdp.h>
+#elif defined(ARDUINO) && (ARDUINO>100)
+#include <NativeEthernet.h>
+#include <NativeEthernetUdp.h>
+#endif
 
 #include "RoveCommManifest.h"
 #include "RoveCommPacket.h"
