+++ src/RoveCommPacket.cpp
@@ -5,8 +5,6 @@
 #include <stdarg.h>
 #include <string.h>
 
-#include "Ethernet.h"
-
 #include "RoveCommPacket.h"
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 namespace roveware
