--- src/srcTeensy/TeensyRoveComm.h
@@ -1,9 +1,0 @@
-#ifndef TeensyRoveComm
-#define TeensyRoveComm
-
-#include "TeensyRoveCommEthernet.h"
-#include "TeensyRoveCommEthernetUdp.h"
-#include "TeensyRoveCommEthernetTCP.h"
-#include "TeensyRoveCommSerial.h"
-
-#endif
